requirements.txt
django-crispy-forms
crispy-bootstrap4
folium
yandex-geocoder
osmnx
networkx
scikit-learn